codes/demand-simulator/data-generator/requirements.txt
# サンプルデータ生成ツール 必要なライブラリ

# YAML設定ファイルの読み込みに使用
PyYAML>=6.0.1
